refactor: 분리되어 있던 온보딩 관련 뷰 합치기 (#66)
