+++ BBANGZIP/BBANGZIP/Sources/Application/BBANGZIPApp.swift
@@ -4,7 +4,7 @@ import SwiftUI
 struct BBANGZIPApp: App {
     var body: some Scene {
         WindowGroup {
-            CustomTabView()
+            OnboardingView()
         }
     }
 }
